+++ omics_data_mgmt/users/backends.py
@@ -6,9 +6,6 @@ class ChariteLDAPBackend(LDAPBackend):
     settings_prefix = 'AUTH_CHARITE_LDAP_'
 
     def authenticate(self, username, password, **kwargs):
-        if len(password) == 0 and not self.settings.PERMIT_EMPTY_PASSWORD:
-            return None
-
         # Ensure username has proper suffix
         if username.find('@') == -1 or username.split('@')[1] != 'CHARITE':
             return None
@@ -25,16 +22,22 @@ class ChariteLDAPBackend(LDAPBackend):
         return username.split('@')[0]
 
 
-# TODO: MDC AD backend
+# MDC AD backend
 class MDCLDAPBackend(LDAPBackend):
     settings_prefix = 'AUTH_MDC_LDAP_'
 
+    def authenticate(self, username, password, **kwargs):
+        # Ensure username has proper suffix
+        if username.find('@') == -1 or username.split('@')[1] != 'MDC-BERLIN':
+            return None
+
+        ldap_user = _LDAPUser(self, username=username.split('@')[0].strip())
+        user = ldap_user.authenticate(password)
+
+        return user
+
     def ldap_to_django_username(self, username):
-        return username
+        return username + '@MDC-BERLIN'
 
     def django_to_ldap_username(self, username):
-        if (username.find('@') == -1 or
-                username.split('@')[1].upper() != 'MDC-BERLIN'):
-            return '**INVALID**' + username
-
         return username.split('@')[0]