+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
     - Changelog in ``CHANGELOG.rst``
     - User profile page (#29)
     - Highlight help link for new users (#30)
+    - Support for multiple LDAP backends (#69)
 - **Adminalerts**
     - Add adminalerts app (#17)
 - **Filesfolders**